+++ CHANGELOG.md
@@ -7,12 +7,14 @@
 
 ### Features
 
+ - Add ability to induce an Input to acquire/store measurements from the web UI
  - Add Input: SHT31 Smart Gadget (Bluetooth) humidity/temperature sensor ([#559](https://github.com/kizniche/mycodo/issues/559))
  - Add blank line to LCD display options ([#579](https://github.com/kizniche/mycodo/issues/579))
 
  ### Miscellaneous
 
  - Add verification for Conditional Statement code
+ - Improve SHT31 Smart Gadget module code
 
 
 ## 7.0.0 (2018-12-08)